+++ Machines/Amiga/Disk.cpp
@@ -234,7 +234,7 @@ uint8_t Chipset::DiskController::get_rdy_trk0_wpro_chng() {
 		(drive.get_is_track_zero() ? 0x10 : 0x00) |
 		(drive.is_read_only() ? 0x08 : 0x00);
 
-	return ~active_high;
+	return uint8_t(~active_high);
 }
 
 void Chipset::DiskController::set_activity_observer(Activity::Observer *observer) {
